further logging improvements
