+++ packages/node-sdk/src/batch-buffer.ts
@@ -72,7 +72,7 @@ export default class BatchBuffer<T> {
     try {
       await this.flushHandler(this.buffer);
 
-      this.logger?.info("flushed buffered items", {
+      this.logger?.debug("flushed buffered items", {
         count: this.buffer.length,
       });
     } catch (error) {
