+++ packages/node-sdk/test/batch-buffer.test.ts
@@ -171,7 +171,7 @@ describe("BatchBuffer", () => {
       await buffer.add("item1");
       await buffer.flush();
 
-      expect(mockLogger.info).toHaveBeenCalledWith("flushed buffered items", {
+      expect(mockLogger.debug).toHaveBeenCalledWith("flushed buffered items", {
         count: 1,
       });
     });
@@ -208,7 +208,7 @@ describe("BatchBuffer", () => {
 
       expect(buffer["timer"]).toBeNull();
 
-      expect(mockLogger.info).toHaveBeenCalledWith("flushed buffered items", {
+      expect(mockLogger.debug).toHaveBeenCalledWith("flushed buffered items", {
         count: 1,
       });
     });
@@ -226,7 +226,7 @@ describe("BatchBuffer", () => {
 
       expect(buffer["timer"]).toBeNull();
 
-      expect(mockLogger.info).toHaveBeenCalledWith("flushed buffered items", {
+      expect(mockLogger.debug).toHaveBeenCalledWith("flushed buffered items", {
         count: 2,
       });
     });
