+++ packages/node-sdk/test/client.test.ts
@@ -7,6 +7,7 @@ import {
   expect,
   it,
   vi,
+  vitest,
 } from "vitest";
 
 import { evaluateTargeting } from "@bucketco/flag-evaluation";
